Parser creating too many ranklist, not linked to players.

+++ server/server/Model/Member/Member.cs
@@ -12,7 +12,7 @@ namespace Server.Model
         private List<MemberRole> _roles = new List<MemberRole>();
 
         public string Name { get; set; }
-        public byte Sex { get; set; }
+        public int Sex { get; set; }
         public int Id { get; set; }
 
         public void AddRole<T>(T Role) where T : MemberRole
